add audio speed control

+++ _third_part/algorithm/audio/com/Algo.Memory.h
@@ -11,7 +11,7 @@
 #define ALGO_CALLOC(count,size)     calloc((void*)count,(i32)size)
 #define ALGO_FREE(block)            free((void*)block)
 #define ALGO_MEM_CPY(dst,src,size)	memcpy((void*)dst,(void*)src,(i32)size)
-#define ALGO_MEM_SET(dst,val,size)	memset((void*)dst,(i32)val,(i32)size)
+#define ALGO_MEM_SET(dst,val,size)	memset((void*)dst,(i8)val,(i32)size)
 #define ALGO_MEM_MOVE(dst,src,size) memmove((void*)dst,(void*)src,(i32)size)
 #else
